pyproject.toml
# Copyright Paul Barker <paul@pbarker.dev>
# SPDX-License-Identifier: Apache-2.0

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pelican = "^4.9.1"
pelican-neighbors = "^1.2.0"
pelican-youtube = "^0.2.1"
python-ulid = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
